// src/app/types/types.ts
  export const userZodSchema = z.object({
    userName: z.string().min(4, "الاسم مطلوب").max(50, "الاسم طويل جداً"),
    email: z.string().email("البريد الإلكتروني غير صحيح"),
    password: z.string().min(4, "يجب إدخال كلمة مرور لا تقل عن 4 أحرف").max(100, "كلمة المرور طويلة جداً"),
    confirmPassword: z.string().min(4, "يجب تأكيد كلمة المرور").max(100, "كلمة المرور طويلة جداً"),
    roleName: z.string(),
  }).refine(data => data.password === data.confirmPassword, {
    message: "كلمات المرور لا تتطابق",
    path: ["confirmPassword"],
    
  });
  
  export type userType = z.infer<typeof userZodSchema>;